Changes to build ctypes_com_server.
The build_comserver() method now accepts an additional, optional
argument 'boot_script'.

The parse_mf_results() method now returns a 3-tuple, the third item is
a sequence containing the names of needed builtin modules.

This is to eventually make it possible to either build single-file
executables or use custom Python builds with a customized set of
builtin modules.  The idea is to create a custom PC\config.c file from
py2exe.